src/requirements.txt
mysql-connector-python
tabulate
python-dotenv
fastapi
uvicorn
email-validator
scikit-learn
pandas
numpy